+++ package.json
@@ -25,6 +25,7 @@
     "@react-navigation/bottom-tabs": "^6.6.1",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/native-stack": "^6.11.0",
+    "@react-navigation/stack": "^6.4.1",
     "@tanstack/react-query": "^5.52.1",
     "expo": "~51.0.28",
     "expo-constants": "~16.0.2",
