+++ src/Screens/Liked.tsx
@@ -1,25 +1,30 @@
-import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import { LikeStackParamList } from "./types";
-import Pokemon from "./Pokemon";
-import LikedList from "./LikedList";
+import PokemonList from "@components/PokemonList";
 
-const Stack = createNativeStackNavigator<LikeStackParamList>();
+import { SafeAreaView } from "react-native-safe-area-context";
+import useLikedPokemons from "../hooks/useLikedPokemons";
+import { TabScreenProps } from "./types";
 
-const Liked = () => {
+type Props = TabScreenProps<"Liked">;
+
+const Liked = ({ navigation }: Props) => {
+  const { pokemonIndex, loading, error, fetchMore, loadingMore } =
+    useLikedPokemons();
+
+  const handlePress = (name: string) => {
+    return navigation.navigate("Pokemon", { name });
+  };
+
   return (
-    <Stack.Navigator>
-      <Stack.Screen
-        options={{ title: "Pokemon", headerShown: false }}
-        name="LikedPokemon"
-        component={LikedList}
+    <SafeAreaView style={{ flex: 1 }}>
+      <PokemonList
+        pokemons={pokemonIndex}
+        loading={loading}
+        error={error}
+        fetchMore={fetchMore}
+        loadingMore={loadingMore}
+        handlePress={handlePress}
       />
-      <Stack.Screen
-        name="Pokemon"
-        component={Pokemon}
-        options={{ title: "Pokemon" }}
-      />
-    </Stack.Navigator>
+    </SafeAreaView>
   );
 };
-
 export default Liked;
