+++ src/Screens/types.tsx
@@ -3,44 +3,29 @@ import {
   CompositeScreenProps,
   NavigatorScreenParams,
 } from "@react-navigation/native";
-import { NativeStackScreenProps } from "@react-navigation/native-stack";
-// import type { StackScreenProps } from "@react-navigation/stack";
+import type { StackScreenProps } from "@react-navigation/stack";
 
-export type PokemonParams = {
-  name: string;
-};
-export type SearchStackParamList = {
-  PokemonIndex: undefined;
-  Pokemon: PokemonParams;
+export type TabStackParamList = {
+  Search: undefined;
+  Liked: undefined;
 };
 
-export type LikeStackParamList = {
-  LikedPokemon: undefined;
-  Pokemon: PokemonParams;
+export type RootStackParamList = {
+  Tab: NavigatorScreenParams<TabStackParamList>;
+  Pokemon: { name: string };
 };
 
-export type RootTabParamsList = {
-  Search: NavigatorScreenParams<SearchStackParamList>;
-  Liked: NavigatorScreenParams<LikeStackParamList>;
-};
+export type RootStackScreenProps<T extends keyof RootStackParamList> =
+  StackScreenProps<RootStackParamList, T>;
 
-export type RootTabScreenProps<T extends keyof RootTabParamsList> =
-  BottomTabScreenProps<RootTabParamsList, T>;
-
-export type SearchScreenProps<T extends keyof SearchStackParamList> =
+export type TabScreenProps<T extends keyof TabStackParamList> =
   CompositeScreenProps<
-    NativeStackScreenProps<SearchStackParamList, T>,
-    RootTabScreenProps<keyof RootTabParamsList>
-  >;
-
-export type LikeScreenProps<T extends keyof LikeStackParamList> =
-  CompositeScreenProps<
-    NativeStackScreenProps<LikeStackParamList, T>,
-    RootTabScreenProps<keyof RootTabParamsList>
+    BottomTabScreenProps<TabStackParamList, T>,
+    RootStackScreenProps<keyof RootStackParamList>
   >;
 
 declare global {
   namespace ReactNavigation {
-    interface RootParamList extends RootTabParamsList {}
+    interface RootParamList extends RootStackParamList {}
   }
 }
